[FIX] compilation warnings
git-svn-id: https://coolfluidsrv.vki.ac.be/svn/coolfluid3/trunk@795 4892e656-8fb9-448a-b5bc-b1919152596f

+++ src/Common/Component.cpp
@@ -488,7 +488,7 @@ std::string Component::tree(Uint level)
 
 /////////////////////////////////////////////////////////////////////////////////////
 
-int Component::get_child_count() const
+size_t Component::get_child_count() const
 {
   return m_components.size();
 }
